feat: Support retrying screenshot capturing for a single client
Enabled via new 'retry' option in config. Use case: test retrying. The idea is to still fetch for all clients in the beginning when createEmail() is called, but only refetch when asked to for clients that need to be retried.

By allowing the api consumer to choose when that retrying happens, we avoid complicated implementation in this project, and can leverage the test runner's ability to retry on failures rather trying to figure this out ourselves